Galois Field Element Class

The `GFElem` class represents an element in a Galois Field (GF). A Galois Field, also known as a finite field, is a field with a finite number of elements. This class provides methods for initializing a field element, checking if it is zero, and displaying its value.

Details

The `GFElem` class has two main fields:

`value`: An integer that represents the value of the Galois field element.

`p`: A prime number that represents the characteristic of the Galois field.

The class provides methods to:

Initialize a Galois Field element with a specified value and characteristic.

Check if the element is zero.

Display the element in a readable format.

Public fields

value

An integer representing the value of the Galois field element.

p

A prime number representing the characteristic of the Galois field.

Methods

Public methods

GFElem$new()

GFElem$iszero()

GFElem$show()

GFElem$clone()

Method new()

Initialize a new GFElem object.

Usage

GFElem$new(value, p)

Arguments

value

An integer. The value of the Galois field element.

p

A prime number. The characteristic of the Galois field.

Returns

A new `GFElem` object.

Method iszero()

Check if the element is zero.

Usage

GFElem$iszero()

Returns

TRUE if the element is zero, otherwise FALSE.

Method show()

Display the element.

Usage

GFElem$show()

Method clone()

The objects of this class are cloneable with this method.

Usage

GFElem$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Create a Galois Field element
elem <- GFElem$new(3, 7)

# Display the Galois Field element
elem$show()
#> GFElem{7}(3)

# Check if the element is zero
elem$iszero()
#> [1] FALSE

# Create a zero element and check
zero_elem <- GFElem$new(0, 7)
zero_elem$iszero()
#> [1] TRUE

# Create an element and check if it is zero
non_zero_elem <- GFElem$new(5, 7)
non_zero_elem$iszero()
#> [1] FALSE

# Display an element
non_zero_elem$show()
#> GFElem{7}(5)