Generate Factors for \(q = 3\)

This function generates factors for a given parameter \(m\) where \(q = 3\). The factors are generated using elements of the Galois Field GF(m-1) and include the point at infinity. The function performs several checks to ensure the validity of the input and uses parallel processing to compute the factors.

Gen3Factors(m)

Arguments

m

An integer representing the parameter for the factors. It must be a multiple of 6, and \(m-1\) must be a prime number.

Value

A list of factors in the Galois Field GF(m-1), including the point at infinity. The result is unlisted and contains the generated factors.

Details

The function checks if \(m\) is a multiple of 6 and if \(m-1\) is a prime number. It then calculates the factors using elements of the Galois Field GF(m-1) and processes them using parallel computation with `future_map` to speed up the operation. The factors include all elements of the field and are calculated using the iterated function approach. The result is unlisted to return a simple list of factors.

Examples

# Generate factors for q = 3
factors <- Gen3Factors(12)
#> Error in (function (.x, .f, ..., .progress = FALSE) {    map_("list", .x, .f, ..., .progress = .progress)})(.x = list(<environment>, <environment>, <environment>, <environment>,     <environment>, <environment>, <environment>, <environment>,     <environment>, <environment>, <environment>, <environment>),     .f = function (...)     {        {            ...furrr_chunk_seeds_i <- ...furrr_chunk_seeds_env[["i"]]            ...furrr_chunk_seeds_env[["i"]] <- ...furrr_chunk_seeds_i +                 1L            assign(x = ".Random.seed", value = ...furrr_chunk_seeds[[...furrr_chunk_seeds_i]],                 envir = globalenv(), inherits = FALSE)        }        NULL        ...furrr_out <- ...furrr_fn(...)        ...furrr_out    }): ℹ In index: 1.
#> Caused by error in `match()`:
#> ! 'match' requires vector arguments
factors
#> Error in eval(expr, envir, enclos): object 'factors' not found