Iterated Function Application

Applies a function iteratively to an initial element until the result repeats. This function generates a sequence of values starting from an initial value, where each subsequent value is obtained by applying a given function to the previous value. The process stops when a repeated value is encountered.

Iterated(f, x)

Arguments

f

A function to apply iteratively. It should take a single argument and return a result of the same type as the input.

x

An initial value to start the iteration. It should be of the same type as the input required by the function `f`.

Value

A list containing the sequence of values obtained by applying the function `f` iteratively to `x`. The sequence ends when a value repeats, and the repeated value is included in the result.

Examples

# Define a function to be iterated
multiply_by_two <- function(x) x * 2

# Apply the function iteratively
result <- Iterated(multiply_by_two, 1)
print(result)
#> [[1]]
#> [1] 1
#> 
#> [[2]]
#> [1] 2
#> 
#> [[3]]
#> [1] 4
#> 
#> [[4]]
#> [1] 8
#> 
#> [[5]]
#> [1] 16
#> 
#> [[6]]
#> [1] 32
#> 
#> [[7]]
#> [1] 64
#> 
#> [[8]]
#> [1] 128
#> 
#> [[9]]
#> [1] 256
#> 
#> [[10]]
#> [1] 512
#> 
#> [[11]]
#> [1] 1024
#> 
#> [[12]]
#> [1] 2048
#> 
#> [[13]]
#> [1] 4096
#> 
#> [[14]]
#> [1] 8192
#> 
#> [[15]]
#> [1] 16384
#> 
#> [[16]]
#> [1] 32768
#> 
#> [[17]]
#> [1] 65536
#> 
#> [[18]]
#> [1] 131072
#> 
#> [[19]]
#> [1] 262144
#> 
#> [[20]]
#> [1] 524288
#> 
#> [[21]]
#> [1] 1048576
#> 
#> [[22]]
#> [1] 2097152
#> 
#> [[23]]
#> [1] 4194304
#> 
#> [[24]]
#> [1] 8388608
#> 
#> [[25]]
#> [1] 16777216
#> 
#> [[26]]
#> [1] 33554432
#> 
#> [[27]]
#> [1] 67108864
#> 
#> [[28]]
#> [1] 134217728
#> 
#> [[29]]
#> [1] 268435456
#> 
#> [[30]]
#> [1] 536870912
#> 
#> [[31]]
#> [1] 1073741824
#> 
#> [[32]]
#> [1] 2147483648
#> 
#> [[33]]
#> [1] 4294967296
#> 
#> [[34]]
#> [1] 8589934592
#> 
#> [[35]]
#> [1] 17179869184
#> 
#> [[36]]
#> [1] 34359738368
#> 
#> [[37]]
#> [1] 68719476736
#> 
#> [[38]]
#> [1] 1.37439e+11
#> 
#> [[39]]
#> [1] 274877906944
#> 
#> [[40]]
#> [1] 549755813888
#> 
#> [[41]]
#> [1] 1.099512e+12
#> 
#> [[42]]
#> [1] 2.199023e+12
#> 
#> [[43]]
#> [1] 4.398047e+12
#> 
#> [[44]]
#> [1] 8.796093e+12
#> 
#> [[45]]
#> [1] 1.759219e+13
#> 
#> [[46]]
#> [1] 3.518437e+13
#> 
#> [[47]]
#> [1] 7.036874e+13
#> 
#> [[48]]
#> [1] 1.407375e+14
#> 
#> [[49]]
#> [1] 2.81475e+14
#> 
#> [[50]]
#> [1] 5.6295e+14
#> 
#> [[51]]
#> [1] 1.1259e+15
#> 
#> [[52]]
#> [1] 2.2518e+15
#> 
#> [[53]]
#> [1] 4.5036e+15
#> 
#> [[54]]
#> [1] 9.007199e+15
#> 
#> [[55]]
#> [1] 1.80144e+16
#> 
#> [[56]]
#> [1] 3.60288e+16
#> 
#> [[57]]
#> [1] 7.205759e+16
#> 
#> [[58]]
#> [1] 1.441152e+17
#> 
#> [[59]]
#> [1] 2.882304e+17
#> 
#> [[60]]
#> [1] 5.764608e+17
#> 
#> [[61]]
#> [1] 1.152922e+18
#> 
#> [[62]]
#> [1] 2.305843e+18
#> 
#> [[63]]
#> [1] 4.611686e+18
#> 
#> [[64]]
#> [1] 9.223372e+18
#> 
#> [[65]]
#> [1] 1.844674e+19
#> 
#> [[66]]
#> [1] 3.689349e+19
#> 
#> [[67]]
#> [1] 7.378698e+19
#> 
#> [[68]]
#> [1] 1.47574e+20
#> 
#> [[69]]
#> [1] 2.951479e+20
#> 
#> [[70]]
#> [1] 5.902958e+20
#> 
#> [[71]]
#> [1] 1.180592e+21
#> 
#> [[72]]
#> [1] 2.361183e+21
#> 
#> [[73]]
#> [1] 4.722366e+21
#> 
#> [[74]]
#> [1] 9.444733e+21
#> 
#> [[75]]
#> [1] 1.888947e+22
#> 
#> [[76]]
#> [1] 3.777893e+22
#> 
#> [[77]]
#> [1] 7.555786e+22
#> 
#> [[78]]
#> [1] 1.511157e+23
#> 
#> [[79]]
#> [1] 3.022315e+23
#> 
#> [[80]]
#> [1] 6.044629e+23
#> 
#> [[81]]
#> [1] 1.208926e+24
#> 
#> [[82]]
#> [1] 2.417852e+24
#> 
#> [[83]]
#> [1] 4.835703e+24
#> 
#> [[84]]
#> [1] 9.671407e+24
#> 
#> [[85]]
#> [1] 1.934281e+25
#> 
#> [[86]]
#> [1] 3.868563e+25
#> 
#> [[87]]
#> [1] 7.737125e+25
#> 
#> [[88]]
#> [1] 1.547425e+26
#> 
#> [[89]]
#> [1] 3.09485e+26
#> 
#> [[90]]
#> [1] 6.1897e+26
#> 
#> [[91]]
#> [1] 1.23794e+27
#> 
#> [[92]]
#> [1] 2.47588e+27
#> 
#> [[93]]
#> [1] 4.95176e+27
#> 
#> [[94]]
#> [1] 9.90352e+27
#> 
#> [[95]]
#> [1] 1.980704e+28
#> 
#> [[96]]
#> [1] 3.961408e+28
#> 
#> [[97]]
#> [1] 7.922816e+28
#> 
#> [[98]]
#> [1] 1.584563e+29
#> 
#> [[99]]
#> [1] 3.169127e+29
#> 
#> [[100]]
#> [1] 6.338253e+29
#> 
#> [[101]]
#> [1] 1.267651e+30
#> 
#> [[102]]
#> [1] 2.535301e+30
#> 
#> [[103]]
#> [1] 5.070602e+30
#> 
#> [[104]]
#> [1] 1.01412e+31
#> 
#> [[105]]
#> [1] 2.028241e+31
#> 
#> [[106]]
#> [1] 4.056482e+31
#> 
#> [[107]]
#> [1] 8.112964e+31
#> 
#> [[108]]
#> [1] 1.622593e+32
#> 
#> [[109]]
#> [1] 3.245186e+32
#> 
#> [[110]]
#> [1] 6.490371e+32
#> 
#> [[111]]
#> [1] 1.298074e+33
#> 
#> [[112]]
#> [1] 2.596148e+33
#> 
#> [[113]]
#> [1] 5.192297e+33
#> 
#> [[114]]
#> [1] 1.038459e+34
#> 
#> [[115]]
#> [1] 2.076919e+34
#> 
#> [[116]]
#> [1] 4.153837e+34
#> 
#> [[117]]
#> [1] 8.307675e+34
#> 
#> [[118]]
#> [1] 1.661535e+35
#> 
#> [[119]]
#> [1] 3.32307e+35
#> 
#> [[120]]
#> [1] 6.64614e+35
#> 
#> [[121]]
#> [1] 1.329228e+36
#> 
#> [[122]]
#> [1] 2.658456e+36
#> 
#> [[123]]
#> [1] 5.316912e+36
#> 
#> [[124]]
#> [1] 1.063382e+37
#> 
#> [[125]]
#> [1] 2.126765e+37
#> 
#> [[126]]
#> [1] 4.25353e+37
#> 
#> [[127]]
#> [1] 8.507059e+37
#> 
#> [[128]]
#> [1] 1.701412e+38
#> 
#> [[129]]
#> [1] 3.402824e+38
#> 
#> [[130]]
#> [1] 6.805647e+38
#> 
#> [[131]]
#> [1] 1.361129e+39
#> 
#> [[132]]
#> [1] 2.722259e+39
#> 
#> [[133]]
#> [1] 5.444518e+39
#> 
#> [[134]]
#> [1] 1.088904e+40
#> 
#> [[135]]
#> [1] 2.177807e+40
#> 
#> [[136]]
#> [1] 4.355614e+40
#> 
#> [[137]]
#> [1] 8.711229e+40
#> 
#> [[138]]
#> [1] 1.742246e+41
#> 
#> [[139]]
#> [1] 3.484491e+41
#> 
#> [[140]]
#> [1] 6.968983e+41
#> 
#> [[141]]
#> [1] 1.393797e+42
#> 
#> [[142]]
#> [1] 2.787593e+42
#> 
#> [[143]]
#> [1] 5.575186e+42
#> 
#> [[144]]
#> [1] 1.115037e+43
#> 
#> [[145]]
#> [1] 2.230075e+43
#> 
#> [[146]]
#> [1] 4.460149e+43
#> 
#> [[147]]
#> [1] 8.920298e+43
#> 
#> [[148]]
#> [1] 1.78406e+44
#> 
#> [[149]]
#> [1] 3.568119e+44
#> 
#> [[150]]
#> [1] 7.136238e+44
#> 
#> [[151]]
#> [1] 1.427248e+45
#> 
#> [[152]]
#> [1] 2.854495e+45
#> 
#> [[153]]
#> [1] 5.708991e+45
#> 
#> [[154]]
#> [1] 1.141798e+46
#> 
#> [[155]]
#> [1] 2.283596e+46
#> 
#> [[156]]
#> [1] 4.567193e+46
#> 
#> [[157]]
#> [1] 9.134385e+46
#> 
#> [[158]]
#> [1] 1.826877e+47
#> 
#> [[159]]
#> [1] 3.653754e+47
#> 
#> [[160]]
#> [1] 7.307508e+47
#> 
#> [[161]]
#> [1] 1.461502e+48
#> 
#> [[162]]
#> [1] 2.923003e+48
#> 
#> [[163]]
#> [1] 5.846007e+48
#> 
#> [[164]]
#> [1] 1.169201e+49
#> 
#> [[165]]
#> [1] 2.338403e+49
#> 
#> [[166]]
#> [1] 4.676805e+49
#> 
#> [[167]]
#> [1] 9.35361e+49
#> 
#> [[168]]
#> [1] 1.870722e+50
#> 
#> [[169]]
#> [1] 3.741444e+50
#> 
#> [[170]]
#> [1] 7.482888e+50
#> 
#> [[171]]
#> [1] 1.496578e+51
#> 
#> [[172]]
#> [1] 2.993155e+51
#> 
#> [[173]]
#> [1] 5.986311e+51
#> 
#> [[174]]
#> [1] 1.197262e+52
#> 
#> [[175]]
#> [1] 2.394524e+52
#> 
#> [[176]]
#> [1] 4.789049e+52
#> 
#> [[177]]
#> [1] 9.578097e+52
#> 
#> [[178]]
#> [1] 1.915619e+53
#> 
#> [[179]]
#> [1] 3.831239e+53
#> 
#> [[180]]
#> [1] 7.662478e+53
#> 
#> [[181]]
#> [1] 1.532496e+54
#> 
#> [[182]]
#> [1] 3.064991e+54
#> 
#> [[183]]
#> [1] 6.129982e+54
#> 
#> [[184]]
#> [1] 1.225996e+55
#> 
#> [[185]]
#> [1] 2.451993e+55
#> 
#> [[186]]
#> [1] 4.903986e+55
#> 
#> [[187]]
#> [1] 9.807971e+55
#> 
#> [[188]]
#> [1] 1.961594e+56
#> 
#> [[189]]
#> [1] 3.923189e+56
#> 
#> [[190]]
#> [1] 7.846377e+56
#> 
#> [[191]]
#> [1] 1.569275e+57
#> 
#> [[192]]
#> [1] 3.138551e+57
#> 
#> [[193]]
#> [1] 6.277102e+57
#> 
#> [[194]]
#> [1] 1.25542e+58
#> 
#> [[195]]
#> [1] 2.510841e+58
#> 
#> [[196]]
#> [1] 5.021681e+58
#> 
#> [[197]]
#> [1] 1.004336e+59
#> 
#> [[198]]
#> [1] 2.008673e+59
#> 
#> [[199]]
#> [1] 4.017345e+59
#> 
#> [[200]]
#> [1] 8.03469e+59
#> 
#> [[201]]
#> [1] 1.606938e+60
#> 
#> [[202]]
#> [1] 3.213876e+60
#> 
#> [[203]]
#> [1] 6.427752e+60
#> 
#> [[204]]
#> [1] 1.28555e+61
#> 
#> [[205]]
#> [1] 2.571101e+61
#> 
#> [[206]]
#> [1] 5.142202e+61
#> 
#> [[207]]
#> [1] 1.02844e+62
#> 
#> [[208]]
#> [1] 2.056881e+62
#> 
#> [[209]]
#> [1] 4.113761e+62
#> 
#> [[210]]
#> [1] 8.227523e+62
#> 
#> [[211]]
#> [1] 1.645505e+63
#> 
#> [[212]]
#> [1] 3.291009e+63
#> 
#> [[213]]
#> [1] 6.582018e+63
#> 
#> [[214]]
#> [1] 1.316404e+64
#> 
#> [[215]]
#> [1] 2.632807e+64
#> 
#> [[216]]
#> [1] 5.265615e+64
#> 
#> [[217]]
#> [1] 1.053123e+65
#> 
#> [[218]]
#> [1] 2.106246e+65
#> 
#> [[219]]
#> [1] 4.212492e+65
#> 
#> [[220]]
#> [1] 8.424983e+65
#> 
#> [[221]]
#> [1] 1.684997e+66
#> 
#> [[222]]
#> [1] 3.369993e+66
#> 
#> [[223]]
#> [1] 6.739987e+66
#> 
#> [[224]]
#> [1] 1.347997e+67
#> 
#> [[225]]
#> [1] 2.695995e+67
#> 
#> [[226]]
#> [1] 5.391989e+67
#> 
#> [[227]]
#> [1] 1.078398e+68
#> 
#> [[228]]
#> [1] 2.156796e+68
#> 
#> [[229]]
#> [1] 4.313591e+68
#> 
#> [[230]]
#> [1] 8.627183e+68
#> 
#> [[231]]
#> [1] 1.725437e+69
#> 
#> [[232]]
#> [1] 3.450873e+69
#> 
#> [[233]]
#> [1] 6.901746e+69
#> 
#> [[234]]
#> [1] 1.380349e+70
#> 
#> [[235]]
#> [1] 2.760699e+70
#> 
#> [[236]]
#> [1] 5.521397e+70
#> 
#> [[237]]
#> [1] 1.104279e+71
#> 
#> [[238]]
#> [1] 2.208559e+71
#> 
#> [[239]]
#> [1] 4.417118e+71
#> 
#> [[240]]
#> [1] 8.834235e+71
#> 
#> [[241]]
#> [1] 1.766847e+72
#> 
#> [[242]]
#> [1] 3.533694e+72
#> 
#> [[243]]
#> [1] 7.067388e+72
#> 
#> [[244]]
#> [1] 1.413478e+73
#> 
#> [[245]]
#> [1] 2.826955e+73
#> 
#> [[246]]
#> [1] 5.653911e+73
#> 
#> [[247]]
#> [1] 1.130782e+74
#> 
#> [[248]]
#> [1] 2.261564e+74
#> 
#> [[249]]
#> [1] 4.523128e+74
#> 
#> [[250]]
#> [1] 9.046257e+74
#> 
#> [[251]]
#> [1] 1.809251e+75
#> 
#> [[252]]
#> [1] 3.618503e+75
#> 
#> [[253]]
#> [1] 7.237006e+75
#> 
#> [[254]]
#> [1] 1.447401e+76
#> 
#> [[255]]
#> [1] 2.894802e+76
#> 
#> [[256]]
#> [1] 5.789604e+76
#> 
#> [[257]]
#> [1] 1.157921e+77
#> 
#> [[258]]
#> [1] 2.315842e+77
#> 
#> [[259]]
#> [1] 4.631684e+77
#> 
#> [[260]]
#> [1] 9.263367e+77
#> 
#> [[261]]
#> [1] 1.852673e+78
#> 
#> [[262]]
#> [1] 3.705347e+78
#> 
#> [[263]]
#> [1] 7.410694e+78
#> 
#> [[264]]
#> [1] 1.482139e+79
#> 
#> [[265]]
#> [1] 2.964277e+79
#> 
#> [[266]]
#> [1] 5.928555e+79
#> 
#> [[267]]
#> [1] 1.185711e+80
#> 
#> [[268]]
#> [1] 2.371422e+80
#> 
#> [[269]]
#> [1] 4.742844e+80
#> 
#> [[270]]
#> [1] 9.485688e+80
#> 
#> [[271]]
#> [1] 1.897138e+81
#> 
#> [[272]]
#> [1] 3.794275e+81
#> 
#> [[273]]
#> [1] 7.58855e+81
#> 
#> [[274]]
#> [1] 1.51771e+82
#> 
#> [[275]]
#> [1] 3.03542e+82
#> 
#> [[276]]
#> [1] 6.07084e+82
#> 
#> [[277]]
#> [1] 1.214168e+83
#> 
#> [[278]]
#> [1] 2.428336e+83
#> 
#> [[279]]
#> [1] 4.856672e+83
#> 
#> [[280]]
#> [1] 9.713344e+83
#> 
#> [[281]]
#> [1] 1.942669e+84
#> 
#> [[282]]
#> [1] 3.885338e+84
#> 
#> [[283]]
#> [1] 7.770676e+84
#> 
#> [[284]]
#> [1] 1.554135e+85
#> 
#> [[285]]
#> [1] 3.10827e+85
#> 
#> [[286]]
#> [1] 6.21654e+85
#> 
#> [[287]]
#> [1] 1.243308e+86
#> 
#> [[288]]
#> [1] 2.486616e+86
#> 
#> [[289]]
#> [1] 4.973232e+86
#> 
#> [[290]]
#> [1] 9.946465e+86
#> 
#> [[291]]
#> [1] 1.989293e+87
#> 
#> [[292]]
#> [1] 3.978586e+87
#> 
#> [[293]]
#> [1] 7.957172e+87
#> 
#> [[294]]
#> [1] 1.591434e+88
#> 
#> [[295]]
#> [1] 3.182869e+88
#> 
#> [[296]]
#> [1] 6.365737e+88
#> 
#> [[297]]
#> [1] 1.273147e+89
#> 
#> [[298]]
#> [1] 2.546295e+89
#> 
#> [[299]]
#> [1] 5.09259e+89
#> 
#> [[300]]
#> [1] 1.018518e+90
#> 
#> [[301]]
#> [1] 2.037036e+90
#> 
#> [[302]]
#> [1] 4.074072e+90
#> 
#> [[303]]
#> [1] 8.148144e+90
#> 
#> [[304]]
#> [1] 1.629629e+91
#> 
#> [[305]]
#> [1] 3.259258e+91
#> 
#> [[306]]
#> [1] 6.518515e+91
#> 
#> [[307]]
#> [1] 1.303703e+92
#> 
#> [[308]]
#> [1] 2.607406e+92
#> 
#> [[309]]
#> [1] 5.214812e+92
#> 
#> [[310]]
#> [1] 1.042962e+93
#> 
#> [[311]]
#> [1] 2.085925e+93
#> 
#> [[312]]
#> [1] 4.17185e+93
#> 
#> [[313]]
#> [1] 8.343699e+93
#> 
#> [[314]]
#> [1] 1.66874e+94
#> 
#> [[315]]
#> [1] 3.33748e+94
#> 
#> [[316]]
#> [1] 6.674959e+94
#> 
#> [[317]]
#> [1] 1.334992e+95
#> 
#> [[318]]
#> [1] 2.669984e+95
#> 
#> [[319]]
#> [1] 5.339968e+95
#> 
#> [[320]]
#> [1] 1.067994e+96
#> 
#> [[321]]
#> [1] 2.135987e+96
#> 
#> [[322]]
#> [1] 4.271974e+96
#> 
#> [[323]]
#> [1] 8.543948e+96
#> 
#> [[324]]
#> [1] 1.70879e+97
#> 
#> [[325]]
#> [1] 3.417579e+97
#> 
#> [[326]]
#> [1] 6.835159e+97
#> 
#> [[327]]
#> [1] 1.367032e+98
#> 
#> [[328]]
#> [1] 2.734063e+98
#> 
#> [[329]]
#> [1] 5.468127e+98
#> 
#> [[330]]
#> [1] 1.093625e+99
#> 
#> [[331]]
#> [1] 2.187251e+99
#> 
#> [[332]]
#> [1] 4.374501e+99
#> 
#> [[333]]
#> [1] 8.749003e+99
#> 
#> [[334]]
#> [1] 1.749801e+100
#> 
#> [[335]]
#> [1] 3.499601e+100
#> 
#> [[336]]
#> [1] 6.999202e+100
#> 
#> [[337]]
#> [1] 1.39984e+101
#> 
#> [[338]]
#> [1] 2.799681e+101
#> 
#> [[339]]
#> [1] 5.599362e+101
#> 
#> [[340]]
#> [1] 1.119872e+102
#> 
#> [[341]]
#> [1] 2.239745e+102
#> 
#> [[342]]
#> [1] 4.479489e+102
#> 
#> [[343]]
#> [1] 8.958979e+102
#> 
#> [[344]]
#> [1] 1.791796e+103
#> 
#> [[345]]
#> [1] 3.583592e+103
#> 
#> [[346]]
#> [1] 7.167183e+103
#> 
#> [[347]]
#> [1] 1.433437e+104
#> 
#> [[348]]
#> [1] 2.866873e+104
#> 
#> [[349]]
#> [1] 5.733747e+104
#> 
#> [[350]]
#> [1] 1.146749e+105
#> 
#> [[351]]
#> [1] 2.293499e+105
#> 
#> [[352]]
#> [1] 4.586997e+105
#> 
#> [[353]]
#> [1] 9.173994e+105
#> 
#> [[354]]
#> [1] 1.834799e+106
#> 
#> [[355]]
#> [1] 3.669598e+106
#> 
#> [[356]]
#> [1] 7.339196e+106
#> 
#> [[357]]
#> [1] 1.467839e+107
#> 
#> [[358]]
#> [1] 2.935678e+107
#> 
#> [[359]]
#> [1] 5.871356e+107
#> 
#> [[360]]
#> [1] 1.174271e+108
#> 
#> [[361]]
#> [1] 2.348543e+108
#> 
#> [[362]]
#> [1] 4.697085e+108
#> 
#> [[363]]
#> [1] 9.39417e+108
#> 
#> [[364]]
#> [1] 1.878834e+109
#> 
#> [[365]]
#> [1] 3.757668e+109
#> 
#> [[366]]
#> [1] 7.515336e+109
#> 
#> [[367]]
#> [1] 1.503067e+110
#> 
#> [[368]]
#> [1] 3.006135e+110
#> 
#> [[369]]
#> [1] 6.012269e+110
#> 
#> [[370]]
#> [1] 1.202454e+111
#> 
#> [[371]]
#> [1] 2.404908e+111
#> 
#> [[372]]
#> [1] 4.809815e+111
#> 
#> [[373]]
#> [1] 9.61963e+111
#> 
#> [[374]]
#> [1] 1.923926e+112
#> 
#> [[375]]
#> [1] 3.847852e+112
#> 
#> [[376]]
#> [1] 7.695704e+112
#> 
#> [[377]]
#> [1] 1.539141e+113
#> 
#> [[378]]
#> [1] 3.078282e+113
#> 
#> [[379]]
#> [1] 6.156563e+113
#> 
#> [[380]]
#> [1] 1.231313e+114
#> 
#> [[381]]
#> [1] 2.462625e+114
#> 
#> [[382]]
#> [1] 4.925251e+114
#> 
#> [[383]]
#> [1] 9.850502e+114
#> 
#> [[384]]
#> [1] 1.9701e+115
#> 
#> [[385]]
#> [1] 3.940201e+115
#> 
#> [[386]]
#> [1] 7.880401e+115
#> 
#> [[387]]
#> [1] 1.57608e+116
#> 
#> [[388]]
#> [1] 3.15216e+116
#> 
#> [[389]]
#> [1] 6.304321e+116
#> 
#> [[390]]
#> [1] 1.260864e+117
#> 
#> [[391]]
#> [1] 2.521728e+117
#> 
#> [[392]]
#> [1] 5.043457e+117
#> 
#> [[393]]
#> [1] 1.008691e+118
#> 
#> [[394]]
#> [1] 2.017383e+118
#> 
#> [[395]]
#> [1] 4.034765e+118
#> 
#> [[396]]
#> [1] 8.069531e+118
#> 
#> [[397]]
#> [1] 1.613906e+119
#> 
#> [[398]]
#> [1] 3.227812e+119
#> 
#> [[399]]
#> [1] 6.455625e+119
#> 
#> [[400]]
#> [1] 1.291125e+120
#> 
#> [[401]]
#> [1] 2.58225e+120
#> 
#> [[402]]
#> [1] 5.1645e+120
#> 
#> [[403]]
#> [1] 1.0329e+121
#> 
#> [[404]]
#> [1] 2.0658e+121
#> 
#> [[405]]
#> [1] 4.1316e+121
#> 
#> [[406]]
#> [1] 8.2632e+121
#> 
#> [[407]]
#> [1] 1.65264e+122
#> 
#> [[408]]
#> [1] 3.30528e+122
#> 
#> [[409]]
#> [1] 6.61056e+122
#> 
#> [[410]]
#> [1] 1.322112e+123
#> 
#> [[411]]
#> [1] 2.644224e+123
#> 
#> [[412]]
#> [1] 5.288448e+123
#> 
#> [[413]]
#> [1] 1.05769e+124
#> 
#> [[414]]
#> [1] 2.115379e+124
#> 
#> [[415]]
#> [1] 4.230758e+124
#> 
#> [[416]]
#> [1] 8.461516e+124
#> 
#> [[417]]
#> [1] 1.692303e+125
#> 
#> [[418]]
#> [1] 3.384607e+125
#> 
#> [[419]]
#> [1] 6.769213e+125
#> 
#> [[420]]
#> [1] 1.353843e+126
#> 
#> [[421]]
#> [1] 2.707685e+126
#> 
#> [[422]]
#> [1] 5.41537e+126
#> 
#> [[423]]
#> [1] 1.083074e+127
#> 
#> [[424]]
#> [1] 2.166148e+127
#> 
#> [[425]]
#> [1] 4.332296e+127
#> 
#> [[426]]
#> [1] 8.664593e+127
#> 
#> [[427]]
#> [1] 1.732919e+128
#> 
#> [[428]]
#> [1] 3.465837e+128
#> 
#> [[429]]
#> [1] 6.931674e+128
#> 
#> [[430]]
#> [1] 1.386335e+129
#> 
#> [[431]]
#> [1] 2.77267e+129
#> 
#> [[432]]
#> [1] 5.545339e+129
#> 
#> [[433]]
#> [1] 1.109068e+130
#> 
#> [[434]]
#> [1] 2.218136e+130
#> 
#> [[435]]
#> [1] 4.436272e+130
#> 
#> [[436]]
#> [1] 8.872543e+130
#> 
#> [[437]]
#> [1] 1.774509e+131
#> 
#> [[438]]
#> [1] 3.549017e+131
#> 
#> [[439]]
#> [1] 7.098034e+131
#> 
#> [[440]]
#> [1] 1.419607e+132
#> 
#> [[441]]
#> [1] 2.839214e+132
#> 
#> [[442]]
#> [1] 5.678428e+132
#> 
#> [[443]]
#> [1] 1.135686e+133
#> 
#> [[444]]
#> [1] 2.271371e+133
#> 
#> [[445]]
#> [1] 4.542742e+133
#> 
#> [[446]]
#> [1] 9.085484e+133
#> 
#> [[447]]
#> [1] 1.817097e+134
#> 
#> [[448]]
#> [1] 3.634194e+134
#> 
#> [[449]]
#> [1] 7.268387e+134
#> 
#> [[450]]
#> [1] 1.453677e+135
#> 
#> [[451]]
#> [1] 2.907355e+135
#> 
#> [[452]]
#> [1] 5.81471e+135
#> 
#> [[453]]
#> [1] 1.162942e+136
#> 
#> [[454]]
#> [1] 2.325884e+136
#> 
#> [[455]]
#> [1] 4.651768e+136
#> 
#> [[456]]
#> [1] 9.303536e+136
#> 
#> [[457]]
#> [1] 1.860707e+137
#> 
#> [[458]]
#> [1] 3.721414e+137
#> 
#> [[459]]
#> [1] 7.442829e+137
#> 
#> [[460]]
#> [1] 1.488566e+138
#> 
#> [[461]]
#> [1] 2.977131e+138
#> 
#> [[462]]
#> [1] 5.954263e+138
#> 
#> [[463]]
#> [1] 1.190853e+139
#> 
#> [[464]]
#> [1] 2.381705e+139
#> 
#> [[465]]
#> [1] 4.76341e+139
#> 
#> [[466]]
#> [1] 9.526821e+139
#> 
#> [[467]]
#> [1] 1.905364e+140
#> 
#> [[468]]
#> [1] 3.810728e+140
#> 
#> [[469]]
#> [1] 7.621456e+140
#> 
#> [[470]]
#> [1] 1.524291e+141
#> 
#> [[471]]
#> [1] 3.048583e+141
#> 
#> [[472]]
#> [1] 6.097165e+141
#> 
#> [[473]]
#> [1] 1.219433e+142
#> 
#> [[474]]
#> [1] 2.438866e+142
#> 
#> [[475]]
#> [1] 4.877732e+142
#> 
#> [[476]]
#> [1] 9.755464e+142
#> 
#> [[477]]
#> [1] 1.951093e+143
#> 
#> [[478]]
#> [1] 3.902186e+143
#> 
#> [[479]]
#> [1] 7.804371e+143
#> 
#> [[480]]
#> [1] 1.560874e+144
#> 
#> [[481]]
#> [1] 3.121749e+144
#> 
#> [[482]]
#> [1] 6.243497e+144
#> 
#> [[483]]
#> [1] 1.248699e+145
#> 
#> [[484]]
#> [1] 2.497399e+145
#> 
#> [[485]]
#> [1] 4.994798e+145
#> 
#> [[486]]
#> [1] 9.989595e+145
#> 
#> [[487]]
#> [1] 1.997919e+146
#> 
#> [[488]]
#> [1] 3.995838e+146
#> 
#> [[489]]
#> [1] 7.991676e+146
#> 
#> [[490]]
#> [1] 1.598335e+147
#> 
#> [[491]]
#> [1] 3.196671e+147
#> 
#> [[492]]
#> [1] 6.393341e+147
#> 
#> [[493]]
#> [1] 1.278668e+148
#> 
#> [[494]]
#> [1] 2.557336e+148
#> 
#> [[495]]
#> [1] 5.114673e+148
#> 
#> [[496]]
#> [1] 1.022935e+149
#> 
#> [[497]]
#> [1] 2.045869e+149
#> 
#> [[498]]
#> [1] 4.091738e+149
#> 
#> [[499]]
#> [1] 8.183477e+149
#> 
#> [[500]]
#> [1] 1.636695e+150
#> 
#> [[501]]
#> [1] 3.273391e+150
#> 
#> [[502]]
#> [1] 6.546781e+150
#> 
#> [[503]]
#> [1] 1.309356e+151
#> 
#> [[504]]
#> [1] 2.618712e+151
#> 
#> [[505]]
#> [1] 5.237425e+151
#> 
#> [[506]]
#> [1] 1.047485e+152
#> 
#> [[507]]
#> [1] 2.09497e+152
#> 
#> [[508]]
#> [1] 4.18994e+152
#> 
#> [[509]]
#> [1] 8.37988e+152
#> 
#> [[510]]
#> [1] 1.675976e+153
#> 
#> [[511]]
#> [1] 3.351952e+153
#> 
#> [[512]]
#> [1] 6.703904e+153
#> 
#> [[513]]
#> [1] 1.340781e+154
#> 
#> [[514]]
#> [1] 2.681562e+154
#> 
#> [[515]]
#> [1] 5.363123e+154
#> 
#> [[516]]
#> [1] 1.072625e+155
#> 
#> [[517]]
#> [1] 2.145249e+155
#> 
#> [[518]]
#> [1] 4.290499e+155
#> 
#> [[519]]
#> [1] 8.580997e+155
#> 
#> [[520]]
#> [1] 1.716199e+156
#> 
#> [[521]]
#> [1] 3.432399e+156
#> 
#> [[522]]
#> [1] 6.864798e+156
#> 
#> [[523]]
#> [1] 1.37296e+157
#> 
#> [[524]]
#> [1] 2.745919e+157
#> 
#> [[525]]
#> [1] 5.491838e+157
#> 
#> [[526]]
#> [1] 1.098368e+158
#> 
#> [[527]]
#> [1] 2.196735e+158
#> 
#> [[528]]
#> [1] 4.393471e+158
#> 
#> [[529]]
#> [1] 8.786941e+158
#> 
#> [[530]]
#> [1] 1.757388e+159
#> 
#> [[531]]
#> [1] 3.514776e+159
#> 
#> [[532]]
#> [1] 7.029553e+159
#> 
#> [[533]]
#> [1] 1.405911e+160
#> 
#> [[534]]
#> [1] 2.811821e+160
#> 
#> [[535]]
#> [1] 5.623642e+160
#> 
#> [[536]]
#> [1] 1.124728e+161
#> 
#> [[537]]
#> [1] 2.249457e+161
#> 
#> [[538]]
#> [1] 4.498914e+161
#> 
#> [[539]]
#> [1] 8.997828e+161
#> 
#> [[540]]
#> [1] 1.799566e+162
#> 
#> [[541]]
#> [1] 3.599131e+162
#> 
#> [[542]]
#> [1] 7.198262e+162
#> 
#> [[543]]
#> [1] 1.439652e+163
#> 
#> [[544]]
#> [1] 2.879305e+163
#> 
#> [[545]]
#> [1] 5.75861e+163
#> 
#> [[546]]
#> [1] 1.151722e+164
#> 
#> [[547]]
#> [1] 2.303444e+164
#> 
#> [[548]]
#> [1] 4.606888e+164
#> 
#> [[549]]
#> [1] 9.213775e+164
#> 
#> [[550]]
#> [1] 1.842755e+165
#> 
#> [[551]]
#> [1] 3.68551e+165
#> 
#> [[552]]
#> [1] 7.37102e+165
#> 
#> [[553]]
#> [1] 1.474204e+166
#> 
#> [[554]]
#> [1] 2.948408e+166
#> 
#> [[555]]
#> [1] 5.896816e+166
#> 
#> [[556]]
#> [1] 1.179363e+167
#> 
#> [[557]]
#> [1] 2.358727e+167
#> 
#> [[558]]
#> [1] 4.717453e+167
#> 
#> [[559]]
#> [1] 9.434906e+167
#> 
#> [[560]]
#> [1] 1.886981e+168
#> 
#> [[561]]
#> [1] 3.773962e+168
#> 
#> [[562]]
#> [1] 7.547925e+168
#> 
#> [[563]]
#> [1] 1.509585e+169
#> 
#> [[564]]
#> [1] 3.01917e+169
#> 
#> [[565]]
#> [1] 6.03834e+169
#> 
#> [[566]]
#> [1] 1.207668e+170
#> 
#> [[567]]
#> [1] 2.415336e+170
#> 
#> [[568]]
#> [1] 4.830672e+170
#> 
#> [[569]]
#> [1] 9.661344e+170
#> 
#> [[570]]
#> [1] 1.932269e+171
#> 
#> [[571]]
#> [1] 3.864538e+171
#> 
#> [[572]]
#> [1] 7.729075e+171
#> 
#> [[573]]
#> [1] 1.545815e+172
#> 
#> [[574]]
#> [1] 3.09163e+172
#> 
#> [[575]]
#> [1] 6.18326e+172
#> 
#> [[576]]
#> [1] 1.236652e+173
#> 
#> [[577]]
#> [1] 2.473304e+173
#> 
#> [[578]]
#> [1] 4.946608e+173
#> 
#> [[579]]
#> [1] 9.893216e+173
#> 
#> [[580]]
#> [1] 1.978643e+174
#> 
#> [[581]]
#> [1] 3.957286e+174
#> 
#> [[582]]
#> [1] 7.914573e+174
#> 
#> [[583]]
#> [1] 1.582915e+175
#> 
#> [[584]]
#> [1] 3.165829e+175
#> 
#> [[585]]
#> [1] 6.331658e+175
#> 
#> [[586]]
#> [1] 1.266332e+176
#> 
#> [[587]]
#> [1] 2.532663e+176
#> 
#> [[588]]
#> [1] 5.065327e+176
#> 
#> [[589]]
#> [1] 1.013065e+177
#> 
#> [[590]]
#> [1] 2.026131e+177
#> 
#> [[591]]
#> [1] 4.052261e+177
#> 
#> [[592]]
#> [1] 8.104523e+177
#> 
#> [[593]]
#> [1] 1.620905e+178
#> 
#> [[594]]
#> [1] 3.241809e+178
#> 
#> [[595]]
#> [1] 6.483618e+178
#> 
#> [[596]]
#> [1] 1.296724e+179
#> 
#> [[597]]
#> [1] 2.593447e+179
#> 
#> [[598]]
#> [1] 5.186894e+179
#> 
#> [[599]]
#> [1] 1.037379e+180
#> 
#> [[600]]
#> [1] 2.074758e+180
#> 
#> [[601]]
#> [1] 4.149516e+180
#> 
#> [[602]]
#> [1] 8.299031e+180
#> 
#> [[603]]
#> [1] 1.659806e+181
#> 
#> [[604]]
#> [1] 3.319612e+181
#> 
#> [[605]]
#> [1] 6.639225e+181
#> 
#> [[606]]
#> [1] 1.327845e+182
#> 
#> [[607]]
#> [1] 2.65569e+182
#> 
#> [[608]]
#> [1] 5.31138e+182
#> 
#> [[609]]
#> [1] 1.062276e+183
#> 
#> [[610]]
#> [1] 2.124552e+183
#> 
#> [[611]]
#> [1] 4.249104e+183
#> 
#> [[612]]
#> [1] 8.498208e+183
#> 
#> [[613]]
#> [1] 1.699642e+184
#> 
#> [[614]]
#> [1] 3.399283e+184
#> 
#> [[615]]
#> [1] 6.798566e+184
#> 
#> [[616]]
#> [1] 1.359713e+185
#> 
#> [[617]]
#> [1] 2.719427e+185
#> 
#> [[618]]
#> [1] 5.438853e+185
#> 
#> [[619]]
#> [1] 1.087771e+186
#> 
#> [[620]]
#> [1] 2.175541e+186
#> 
#> [[621]]
#> [1] 4.351082e+186
#> 
#> [[622]]
#> [1] 8.702165e+186
#> 
#> [[623]]
#> [1] 1.740433e+187
#> 
#> [[624]]
#> [1] 3.480866e+187
#> 
#> [[625]]
#> [1] 6.961732e+187
#> 
#> [[626]]
#> [1] 1.392346e+188
#> 
#> [[627]]
#> [1] 2.784693e+188
#> 
#> [[628]]
#> [1] 5.569386e+188
#> 
#> [[629]]
#> [1] 1.113877e+189
#> 
#> [[630]]
#> [1] 2.227754e+189
#> 
#> [[631]]
#> [1] 4.455508e+189
#> 
#> [[632]]
#> [1] 8.911017e+189
#> 
#> [[633]]
#> [1] 1.782203e+190
#> 
#> [[634]]
#> [1] 3.564407e+190
#> 
#> [[635]]
#> [1] 7.128813e+190
#> 
#> [[636]]
#> [1] 1.425763e+191
#> 
#> [[637]]
#> [1] 2.851525e+191
#> 
#> [[638]]
#> [1] 5.703051e+191
#> 
#> [[639]]
#> [1] 1.14061e+192
#> 
#> [[640]]
#> [1] 2.28122e+192
#> 
#> [[641]]
#> [1] 4.562441e+192
#> 
#> [[642]]
#> [1] 9.124881e+192
#> 
#> [[643]]
#> [1] 1.824976e+193
#> 
#> [[644]]
#> [1] 3.649952e+193
#> 
#> [[645]]
#> [1] 7.299905e+193
#> 
#> [[646]]
#> [1] 1.459981e+194
#> 
#> [[647]]
#> [1] 2.919962e+194
#> 
#> [[648]]
#> [1] 5.839924e+194
#> 
#> [[649]]
#> [1] 1.167985e+195
#> 
#> [[650]]
#> [1] 2.33597e+195
#> 
#> [[651]]
#> [1] 4.671939e+195
#> 
#> [[652]]
#> [1] 9.343878e+195
#> 
#> [[653]]
#> [1] 1.868776e+196
#> 
#> [[654]]
#> [1] 3.737551e+196
#> 
#> [[655]]
#> [1] 7.475103e+196
#> 
#> [[656]]
#> [1] 1.495021e+197
#> 
#> [[657]]
#> [1] 2.990041e+197
#> 
#> [[658]]
#> [1] 5.980082e+197
#> 
#> [[659]]
#> [1] 1.196016e+198
#> 
#> [[660]]
#> [1] 2.392033e+198
#> 
#> [[661]]
#> [1] 4.784066e+198
#> 
#> [[662]]
#> [1] 9.568131e+198
#> 
#> [[663]]
#> [1] 1.913626e+199
#> 
#> [[664]]
#> [1] 3.827253e+199
#> 
#> [[665]]
#> [1] 7.654505e+199
#> 
#> [[666]]
#> [1] 1.530901e+200
#> 
#> [[667]]
#> [1] 3.061802e+200
#> 
#> [[668]]
#> [1] 6.123604e+200
#> 
#> [[669]]
#> [1] 1.224721e+201
#> 
#> [[670]]
#> [1] 2.449442e+201
#> 
#> [[671]]
#> [1] 4.898883e+201
#> 
#> [[672]]
#> [1] 9.797767e+201
#> 
#> [[673]]
#> [1] 1.959553e+202
#> 
#> [[674]]
#> [1] 3.919107e+202
#> 
#> [[675]]
#> [1] 7.838213e+202
#> 
#> [[676]]
#> [1] 1.567643e+203
#> 
#> [[677]]
#> [1] 3.135285e+203
#> 
#> [[678]]
#> [1] 6.270571e+203
#> 
#> [[679]]
#> [1] 1.254114e+204
#> 
#> [[680]]
#> [1] 2.508228e+204
#> 
#> [[681]]
#> [1] 5.016457e+204
#> 
#> [[682]]
#> [1] 1.003291e+205
#> 
#> [[683]]
#> [1] 2.006583e+205
#> 
#> [[684]]
#> [1] 4.013165e+205
#> 
#> [[685]]
#> [1] 8.02633e+205
#> 
#> [[686]]
#> [1] 1.605266e+206
#> 
#> [[687]]
#> [1] 3.210532e+206
#> 
#> [[688]]
#> [1] 6.421064e+206
#> 
#> [[689]]
#> [1] 1.284213e+207
#> 
#> [[690]]
#> [1] 2.568426e+207
#> 
#> [[691]]
#> [1] 5.136851e+207
#> 
#> [[692]]
#> [1] 1.02737e+208
#> 
#> [[693]]
#> [1] 2.054741e+208
#> 
#> [[694]]
#> [1] 4.109481e+208
#> 
#> [[695]]
#> [1] 8.218962e+208
#> 
#> [[696]]
#> [1] 1.643792e+209
#> 
#> [[697]]
#> [1] 3.287585e+209
#> 
#> [[698]]
#> [1] 6.57517e+209
#> 
#> [[699]]
#> [1] 1.315034e+210
#> 
#> [[700]]
#> [1] 2.630068e+210
#> 
#> [[701]]
#> [1] 5.260136e+210
#> 
#> [[702]]
#> [1] 1.052027e+211
#> 
#> [[703]]
#> [1] 2.104054e+211
#> 
#> [[704]]
#> [1] 4.208109e+211
#> 
#> [[705]]
#> [1] 8.416217e+211
#> 
#> [[706]]
#> [1] 1.683243e+212
#> 
#> [[707]]
#> [1] 3.366487e+212
#> 
#> [[708]]
#> [1] 6.732974e+212
#> 
#> [[709]]
#> [1] 1.346595e+213
#> 
#> [[710]]
#> [1] 2.69319e+213
#> 
#> [[711]]
#> [1] 5.386379e+213
#> 
#> [[712]]
#> [1] 1.077276e+214
#> 
#> [[713]]
#> [1] 2.154552e+214
#> 
#> [[714]]
#> [1] 4.309103e+214
#> 
#> [[715]]
#> [1] 8.618207e+214
#> 
#> [[716]]
#> [1] 1.723641e+215
#> 
#> [[717]]
#> [1] 3.447283e+215
#> 
#> [[718]]
#> [1] 6.894565e+215
#> 
#> [[719]]
#> [1] 1.378913e+216
#> 
#> [[720]]
#> [1] 2.757826e+216
#> 
#> [[721]]
#> [1] 5.515652e+216
#> 
#> [[722]]
#> [1] 1.10313e+217
#> 
#> [[723]]
#> [1] 2.206261e+217
#> 
#> [[724]]
#> [1] 4.412522e+217
#> 
#> [[725]]
#> [1] 8.825044e+217
#> 
#> [[726]]
#> [1] 1.765009e+218
#> 
#> [[727]]
#> [1] 3.530017e+218
#> 
#> [[728]]
#> [1] 7.060035e+218
#> 
#> [[729]]
#> [1] 1.412007e+219
#> 
#> [[730]]
#> [1] 2.824014e+219
#> 
#> [[731]]
#> [1] 5.648028e+219
#> 
#> [[732]]
#> [1] 1.129606e+220
#> 
#> [[733]]
#> [1] 2.259211e+220
#> 
#> [[734]]
#> [1] 4.518422e+220
#> 
#> [[735]]
#> [1] 9.036845e+220
#> 
#> [[736]]
#> [1] 1.807369e+221
#> 
#> [[737]]
#> [1] 3.614738e+221
#> 
#> [[738]]
#> [1] 7.229476e+221
#> 
#> [[739]]
#> [1] 1.445895e+222
#> 
#> [[740]]
#> [1] 2.89179e+222
#> 
#> [[741]]
#> [1] 5.783581e+222
#> 
#> [[742]]
#> [1] 1.156716e+223
#> 
#> [[743]]
#> [1] 2.313432e+223
#> 
#> [[744]]
#> [1] 4.626864e+223
#> 
#> [[745]]
#> [1] 9.253729e+223
#> 
#> [[746]]
#> [1] 1.850746e+224
#> 
#> [[747]]
#> [1] 3.701492e+224
#> 
#> [[748]]
#> [1] 7.402983e+224
#> 
#> [[749]]
#> [1] 1.480597e+225
#> 
#> [[750]]
#> [1] 2.961193e+225
#> 
#> [[751]]
#> [1] 5.922387e+225
#> 
#> [[752]]
#> [1] 1.184477e+226
#> 
#> [[753]]
#> [1] 2.368955e+226
#> 
#> [[754]]
#> [1] 4.737909e+226
#> 
#> [[755]]
#> [1] 9.475818e+226
#> 
#> [[756]]
#> [1] 1.895164e+227
#> 
#> [[757]]
#> [1] 3.790327e+227
#> 
#> [[758]]
#> [1] 7.580655e+227
#> 
#> [[759]]
#> [1] 1.516131e+228
#> 
#> [[760]]
#> [1] 3.032262e+228
#> 
#> [[761]]
#> [1] 6.064524e+228
#> 
#> [[762]]
#> [1] 1.212905e+229
#> 
#> [[763]]
#> [1] 2.42581e+229
#> 
#> [[764]]
#> [1] 4.851619e+229
#> 
#> [[765]]
#> [1] 9.703238e+229
#> 
#> [[766]]
#> [1] 1.940648e+230
#> 
#> [[767]]
#> [1] 3.881295e+230
#> 
#> [[768]]
#> [1] 7.76259e+230
#> 
#> [[769]]
#> [1] 1.552518e+231
#> 
#> [[770]]
#> [1] 3.105036e+231
#> 
#> [[771]]
#> [1] 6.210072e+231
#> 
#> [[772]]
#> [1] 1.242014e+232
#> 
#> [[773]]
#> [1] 2.484029e+232
#> 
#> [[774]]
#> [1] 4.968058e+232
#> 
#> [[775]]
#> [1] 9.936116e+232
#> 
#> [[776]]
#> [1] 1.987223e+233
#> 
#> [[777]]
#> [1] 3.974446e+233
#> 
#> [[778]]
#> [1] 7.948893e+233
#> 
#> [[779]]
#> [1] 1.589779e+234
#> 
#> [[780]]
#> [1] 3.179557e+234
#> 
#> [[781]]
#> [1] 6.359114e+234
#> 
#> [[782]]
#> [1] 1.271823e+235
#> 
#> [[783]]
#> [1] 2.543646e+235
#> 
#> [[784]]
#> [1] 5.087291e+235
#> 
#> [[785]]
#> [1] 1.017458e+236
#> 
#> [[786]]
#> [1] 2.034917e+236
#> 
#> [[787]]
#> [1] 4.069833e+236
#> 
#> [[788]]
#> [1] 8.139666e+236
#> 
#> [[789]]
#> [1] 1.627933e+237
#> 
#> [[790]]
#> [1] 3.255866e+237
#> 
#> [[791]]
#> [1] 6.511733e+237
#> 
#> [[792]]
#> [1] 1.302347e+238
#> 
#> [[793]]
#> [1] 2.604693e+238
#> 
#> [[794]]
#> [1] 5.209386e+238
#> 
#> [[795]]
#> [1] 1.041877e+239
#> 
#> [[796]]
#> [1] 2.083755e+239
#> 
#> [[797]]
#> [1] 4.167509e+239
#> 
#> [[798]]
#> [1] 8.335018e+239
#> 
#> [[799]]
#> [1] 1.667004e+240
#> 
#> [[800]]
#> [1] 3.334007e+240
#> 
#> [[801]]
#> [1] 6.668014e+240
#> 
#> [[802]]
#> [1] 1.333603e+241
#> 
#> [[803]]
#> [1] 2.667206e+241
#> 
#> [[804]]
#> [1] 5.334412e+241
#> 
#> [[805]]
#> [1] 1.066882e+242
#> 
#> [[806]]
#> [1] 2.133765e+242
#> 
#> [[807]]
#> [1] 4.267529e+242
#> 
#> [[808]]
#> [1] 8.535058e+242
#> 
#> [[809]]
#> [1] 1.707012e+243
#> 
#> [[810]]
#> [1] 3.414023e+243
#> 
#> [[811]]
#> [1] 6.828047e+243
#> 
#> [[812]]
#> [1] 1.365609e+244
#> 
#> [[813]]
#> [1] 2.731219e+244
#> 
#> [[814]]
#> [1] 5.462437e+244
#> 
#> [[815]]
#> [1] 1.092487e+245
#> 
#> [[816]]
#> [1] 2.184975e+245
#> 
#> [[817]]
#> [1] 4.36995e+245
#> 
#> [[818]]
#> [1] 8.7399e+245
#> 
#> [[819]]
#> [1] 1.74798e+246
#> 
#> [[820]]
#> [1] 3.49596e+246
#> 
#> [[821]]
#> [1] 6.99192e+246
#> 
#> [[822]]
#> [1] 1.398384e+247
#> 
#> [[823]]
#> [1] 2.796768e+247
#> 
#> [[824]]
#> [1] 5.593536e+247
#> 
#> [[825]]
#> [1] 1.118707e+248
#> 
#> [[826]]
#> [1] 2.237414e+248
#> 
#> [[827]]
#> [1] 4.474829e+248
#> 
#> [[828]]
#> [1] 8.949657e+248
#> 
#> [[829]]
#> [1] 1.789931e+249
#> 
#> [[830]]
#> [1] 3.579863e+249
#> 
#> [[831]]
#> [1] 7.159726e+249
#> 
#> [[832]]
#> [1] 1.431945e+250
#> 
#> [[833]]
#> [1] 2.86389e+250
#> 
#> [[834]]
#> [1] 5.727781e+250
#> 
#> [[835]]
#> [1] 1.145556e+251
#> 
#> [[836]]
#> [1] 2.291112e+251
#> 
#> [[837]]
#> [1] 4.582225e+251
#> 
#> [[838]]
#> [1] 9.164449e+251
#> 
#> [[839]]
#> [1] 1.83289e+252
#> 
#> [[840]]
#> [1] 3.66578e+252
#> 
#> [[841]]
#> [1] 7.331559e+252
#> 
#> [[842]]
#> [1] 1.466312e+253
#> 
#> [[843]]
#> [1] 2.932624e+253
#> 
#> [[844]]
#> [1] 5.865248e+253
#> 
#> [[845]]
#> [1] 1.17305e+254
#> 
#> [[846]]
#> [1] 2.346099e+254
#> 
#> [[847]]
#> [1] 4.692198e+254
#> 
#> [[848]]
#> [1] 9.384396e+254
#> 
#> [[849]]
#> [1] 1.876879e+255
#> 
#> [[850]]
#> [1] 3.753758e+255
#> 
#> [[851]]
#> [1] 7.507517e+255
#> 
#> [[852]]
#> [1] 1.501503e+256
#> 
#> [[853]]
#> [1] 3.003007e+256
#> 
#> [[854]]
#> [1] 6.006013e+256
#> 
#> [[855]]
#> [1] 1.201203e+257
#> 
#> [[856]]
#> [1] 2.402405e+257
#> 
#> [[857]]
#> [1] 4.804811e+257
#> 
#> [[858]]
#> [1] 9.609622e+257
#> 
#> [[859]]
#> [1] 1.921924e+258
#> 
#> [[860]]
#> [1] 3.843849e+258
#> 
#> [[861]]
#> [1] 7.687697e+258
#> 
#> [[862]]
#> [1] 1.537539e+259
#> 
#> [[863]]
#> [1] 3.075079e+259
#> 
#> [[864]]
#> [1] 6.150158e+259
#> 
#> [[865]]
#> [1] 1.230032e+260
#> 
#> [[866]]
#> [1] 2.460063e+260
#> 
#> [[867]]
#> [1] 4.920126e+260
#> 
#> [[868]]
#> [1] 9.840252e+260
#> 
#> [[869]]
#> [1] 1.96805e+261
#> 
#> [[870]]
#> [1] 3.936101e+261
#> 
#> [[871]]
#> [1] 7.872202e+261
#> 
#> [[872]]
#> [1] 1.57444e+262
#> 
#> [[873]]
#> [1] 3.148881e+262
#> 
#> [[874]]
#> [1] 6.297762e+262
#> 
#> [[875]]
#> [1] 1.259552e+263
#> 
#> [[876]]
#> [1] 2.519105e+263
#> 
#> [[877]]
#> [1] 5.038209e+263
#> 
#> [[878]]
#> [1] 1.007642e+264
#> 
#> [[879]]
#> [1] 2.015284e+264
#> 
#> [[880]]
#> [1] 4.030567e+264
#> 
#> [[881]]
#> [1] 8.061135e+264
#> 
#> [[882]]
#> [1] 1.612227e+265
#> 
#> [[883]]
#> [1] 3.224454e+265
#> 
#> [[884]]
#> [1] 6.448908e+265
#> 
#> [[885]]
#> [1] 1.289782e+266
#> 
#> [[886]]
#> [1] 2.579563e+266
#> 
#> [[887]]
#> [1] 5.159126e+266
#> 
#> [[888]]
#> [1] 1.031825e+267
#> 
#> [[889]]
#> [1] 2.063651e+267
#> 
#> [[890]]
#> [1] 4.127301e+267
#> 
#> [[891]]
#> [1] 8.254602e+267
#> 
#> [[892]]
#> [1] 1.65092e+268
#> 
#> [[893]]
#> [1] 3.301841e+268
#> 
#> [[894]]
#> [1] 6.603682e+268
#> 
#> [[895]]
#> [1] 1.320736e+269
#> 
#> [[896]]
#> [1] 2.641473e+269
#> 
#> [[897]]
#> [1] 5.282945e+269
#> 
#> [[898]]
#> [1] 1.056589e+270
#> 
#> [[899]]
#> [1] 2.113178e+270
#> 
#> [[900]]
#> [1] 4.226356e+270
#> 
#> [[901]]
#> [1] 8.452712e+270
#> 
#> [[902]]
#> [1] 1.690542e+271
#> 
#> [[903]]
#> [1] 3.381085e+271
#> 
#> [[904]]
#> [1] 6.76217e+271
#> 
#> [[905]]
#> [1] 1.352434e+272
#> 
#> [[906]]
#> [1] 2.704868e+272
#> 
#> [[907]]
#> [1] 5.409736e+272
#> 
#> [[908]]
#> [1] 1.081947e+273
#> 
#> [[909]]
#> [1] 2.163894e+273
#> 
#> [[910]]
#> [1] 4.327789e+273
#> 
#> [[911]]
#> [1] 8.655578e+273
#> 
#> [[912]]
#> [1] 1.731116e+274
#> 
#> [[913]]
#> [1] 3.462231e+274
#> 
#> [[914]]
#> [1] 6.924462e+274
#> 
#> [[915]]
#> [1] 1.384892e+275
#> 
#> [[916]]
#> [1] 2.769785e+275
#> 
#> [[917]]
#> [1] 5.53957e+275
#> 
#> [[918]]
#> [1] 1.107914e+276
#> 
#> [[919]]
#> [1] 2.215828e+276
#> 
#> [[920]]
#> [1] 4.431656e+276
#> 
#> [[921]]
#> [1] 8.863311e+276
#> 
#> [[922]]
#> [1] 1.772662e+277
#> 
#> [[923]]
#> [1] 3.545325e+277
#> 
#> [[924]]
#> [1] 7.090649e+277
#> 
#> [[925]]
#> [1] 1.41813e+278
#> 
#> [[926]]
#> [1] 2.83626e+278
#> 
#> [[927]]
#> [1] 5.672519e+278
#> 
#> [[928]]
#> [1] 1.134504e+279
#> 
#> [[929]]
#> [1] 2.269008e+279
#> 
#> [[930]]
#> [1] 4.538015e+279
#> 
#> [[931]]
#> [1] 9.076031e+279
#> 
#> [[932]]
#> [1] 1.815206e+280
#> 
#> [[933]]
#> [1] 3.630412e+280
#> 
#> [[934]]
#> [1] 7.260825e+280
#> 
#> [[935]]
#> [1] 1.452165e+281
#> 
#> [[936]]
#> [1] 2.90433e+281
#> 
#> [[937]]
#> [1] 5.80866e+281
#> 
#> [[938]]
#> [1] 1.161732e+282
#> 
#> [[939]]
#> [1] 2.323464e+282
#> 
#> [[940]]
#> [1] 4.646928e+282
#> 
#> [[941]]
#> [1] 9.293856e+282
#> 
#> [[942]]
#> [1] 1.858771e+283
#> 
#> [[943]]
#> [1] 3.717542e+283
#> 
#> [[944]]
#> [1] 7.435085e+283
#> 
#> [[945]]
#> [1] 1.487017e+284
#> 
#> [[946]]
#> [1] 2.974034e+284
#> 
#> [[947]]
#> [1] 5.948068e+284
#> 
#> [[948]]
#> [1] 1.189614e+285
#> 
#> [[949]]
#> [1] 2.379227e+285
#> 
#> [[950]]
#> [1] 4.758454e+285
#> 
#> [[951]]
#> [1] 9.516908e+285
#> 
#> [[952]]
#> [1] 1.903382e+286
#> 
#> [[953]]
#> [1] 3.806763e+286
#> 
#> [[954]]
#> [1] 7.613527e+286
#> 
#> [[955]]
#> [1] 1.522705e+287
#> 
#> [[956]]
#> [1] 3.045411e+287
#> 
#> [[957]]
#> [1] 6.090821e+287
#> 
#> [[958]]
#> [1] 1.218164e+288
#> 
#> [[959]]
#> [1] 2.436329e+288
#> 
#> [[960]]
#> [1] 4.872657e+288
#> 
#> [[961]]
#> [1] 9.745314e+288
#> 
#> [[962]]
#> [1] 1.949063e+289
#> 
#> [[963]]
#> [1] 3.898126e+289
#> 
#> [[964]]
#> [1] 7.796251e+289
#> 
#> [[965]]
#> [1] 1.55925e+290
#> 
#> [[966]]
#> [1] 3.1185e+290
#> 
#> [[967]]
#> [1] 6.237001e+290
#> 
#> [[968]]
#> [1] 1.2474e+291
#> 
#> [[969]]
#> [1] 2.4948e+291
#> 
#> [[970]]
#> [1] 4.989601e+291
#> 
#> [[971]]
#> [1] 9.979202e+291
#> 
#> [[972]]
#> [1] 1.99584e+292
#> 
#> [[973]]
#> [1] 3.991681e+292
#> 
#> [[974]]
#> [1] 7.983361e+292
#> 
#> [[975]]
#> [1] 1.596672e+293
#> 
#> [[976]]
#> [1] 3.193344e+293
#> 
#> [[977]]
#> [1] 6.386689e+293
#> 
#> [[978]]
#> [1] 1.277338e+294
#> 
#> [[979]]
#> [1] 2.554676e+294
#> 
#> [[980]]
#> [1] 5.109351e+294
#> 
#> [[981]]
#> [1] 1.02187e+295
#> 
#> [[982]]
#> [1] 2.04374e+295
#> 
#> [[983]]
#> [1] 4.087481e+295
#> 
#> [[984]]
#> [1] 8.174962e+295
#> 
#> [[985]]
#> [1] 1.634992e+296
#> 
#> [[986]]
#> [1] 3.269985e+296
#> 
#> [[987]]
#> [1] 6.53997e+296
#> 
#> [[988]]
#> [1] 1.307994e+297
#> 
#> [[989]]
#> [1] 2.615988e+297
#> 
#> [[990]]
#> [1] 5.231976e+297
#> 
#> [[991]]
#> [1] 1.046395e+298
#> 
#> [[992]]
#> [1] 2.09279e+298
#> 
#> [[993]]
#> [1] 4.18558e+298
#> 
#> [[994]]
#> [1] 8.371161e+298
#> 
#> [[995]]
#> [1] 1.674232e+299
#> 
#> [[996]]
#> [1] 3.348464e+299
#> 
#> [[997]]
#> [1] 6.696929e+299
#> 
#> [[998]]
#> [1] 1.339386e+300
#> 
#> [[999]]
#> [1] 2.678772e+300
#> 
#> [[1000]]
#> [1] 5.357543e+300
#> 
#> [[1001]]
#> [1] 1.071509e+301
#> 
#> [[1002]]
#> [1] 2.143017e+301
#> 
#> [[1003]]
#> [1] 4.286034e+301
#> 
#> [[1004]]
#> [1] 8.572069e+301
#> 
#> [[1005]]
#> [1] 1.714414e+302
#> 
#> [[1006]]
#> [1] 3.428828e+302
#> 
#> [[1007]]
#> [1] 6.857655e+302
#> 
#> [[1008]]
#> [1] 1.371531e+303
#> 
#> [[1009]]
#> [1] 2.743062e+303
#> 
#> [[1010]]
#> [1] 5.486124e+303
#> 
#> [[1011]]
#> [1] 1.097225e+304
#> 
#> [[1012]]
#> [1] 2.19445e+304
#> 
#> [[1013]]
#> [1] 4.388899e+304
#> 
#> [[1014]]
#> [1] 8.777799e+304
#> 
#> [[1015]]
#> [1] 1.75556e+305
#> 
#> [[1016]]
#> [1] 3.511119e+305
#> 
#> [[1017]]
#> [1] 7.022239e+305
#> 
#> [[1018]]
#> [1] 1.404448e+306
#> 
#> [[1019]]
#> [1] 2.808896e+306
#> 
#> [[1020]]
#> [1] 5.617791e+306
#> 
#> [[1021]]
#> [1] 1.123558e+307
#> 
#> [[1022]]
#> [1] 2.247116e+307
#> 
#> [[1023]]
#> [1] 4.494233e+307
#> 
#> [[1024]]
#> [1] 8.988466e+307
#> 
#> [[1025]]
#> [1] Inf
#> 

# Define a GFElem element and apply a function to it
elem <- GFElem$new(3, 7)
iterated <- Iterated(function(x) Multiply(x, elem), elem)
#> Error in match(x, table, nomatch = 0L): 'match' requires vector arguments
iterated
#> Error in eval(expr, envir, enclos): object 'iterated' not found